Oaks Crisis RPA (#5)
* oaks-crisis: init commit

* oaks-crisis: Custom report automation finished

* oaks-crisis: Finish crisis reports downloads, fix to_dates in appleseed and abhs

* oaks-crisis: Ready for linux deployment

* oaks-crisis: add crisis excel file

* oaks-crisis: add oaks-crisis automation to 0_14_15_r_r.sh

* oaks-crisis: Linux migration finished

* oaks-crisis: reset xlsx for rerunning for prev months

* Delete crisis_sfy_2021.xlsx

* oaks-crisis: re-adding xlsx

* Delete crisis_sfy_2021.xlsx

* oaks-crisis: retrying xlsx file...

* oaks-crisis: Ran Crisis RPA for every month since July 2020

* fremont_isl: Ensure Fremont ISL RPA doesn't delete crisis spreadsheet

* oaks-crisis: Implement schedule and row 21 add-on

* oaks-crisis: Fix dumb changes I made to other the Appleseed RPAs

+++ src/automation/fremont_isl.py
@@ -466,7 +466,8 @@ def fremont_isl(from_date):
                  "Reports shared drive: https://drive.google.com/drive/folders/1lYsW4yfourbnFYJB3GLh6br7D1_3LOcd" \
                  % from_date.strftime('%m-%d-%Y')
     for filename in os.listdir('src/csv'):
-        os.remove('src/csv/%s' % filename)
+        if not filename.endswith('.xlsx'):
+            os.remove('src/csv/%s' % filename)
     for filename in os.listdir('src/pdf'):
         os.remove('src/pdf/%s' % filename)
     shutil.rmtree(folder_path)
